feature/86cuvrwhr-sign-in-authentication: updates to UI and window navigation.
The graphic on the left-hand side of the screen has been added and referenced correctly.

Artist accounts are now navigated to the correct page when signed in.
The logic is there for bride accounts to be navigated to the correct page, however the window reference is still yet to be confirmed.